requirements.txt
# Core dependencies
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2

# API interaction
requests==2.31.0
python-dotenv==1.0.0

# Data analysis and linear algebra
scipy==1.11.3
scikit-learn==1.3.0

# Optional - for interactive development
jupyter==1.0.0
ipykernel==6.25.1

# Optional - for advanced visualizations
plotly==5.16.1